refactor: move 'ToggleManager' to 'services' package
